requirements.txt
sqlalchemy
psycopg2-binary
antropic
python-dotenv
pandas